backend/pyproject.toml
[project]
name = "quorum"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Rob Cowie", email = "rob_cowie@mac.com" }]
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.7",
    "itsdangerous>=2.2.0",
    "pydantic-settings>=2.7.1",
]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

[dependency-groups]
dev = ["ipython>=8.31.0"]
